Fixed syntax errors
Our original implementation of the selection and case stmts had
the wrong bash syntax. We have simply reverted to using regular
case statements. This is not optimal, because we would like to
have error trapping in the case where the user doesn't enter
'y' or 'n'.

Also smtp-cli doesn't seem to like spanning arguments over several
lines using the '\' backslash character, so we just had to leave the
178-character invocation of smtp-cli alone.

TO DO:

- wrap the bash menu in some kind of loop for error handling
- try to implement a $BODY variable (had problems with this
  because gmail stopped reading in input after the first SPC
  char was detected)