Tweaked the script to accept a parameter for number of pomodoros
Before this value was hard-coded to 1, but since other scripts call
pomo2beeminder.sh I think it's better to allow this value to be passed
in as a param. Also the Beeminder Bot requires comments to be
double-quoted, so I had to escape the double quotes, i.e. \" because
I am using double-quotes already so that shell expansion for $1
parameter will work.

You know that if you are using double quotes, a pair will close the
expression; that is why I have to escape the two double quotes around
the comment to the Beeminder Bot.